+++ Sources/MapboxSearch/PublicAPI/Use Cases/Address Autofill/Models/Suggestion/AddressAutofill.Suggestion+SearchResult.swift
@@ -24,6 +24,7 @@ extension AddressAutofill.Suggestion {
         }
 
         return .init(
+            name: searchResult.name,
             formattedAddress: formattedAddress,
             coordinate: searchResult.coordinate,
             addressComponents: try address.toAutofillComponents()