+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Updated
+- [Autofill]: `name` field exposed for the `AddressAutofill.Suggestion`.
+- [Autofill]: exposed query requirements constant in `AddressAutofill.Query.Requirements`.
+- [Autofill]: added example to the Demo project
+
 ## [1.0.0-beta.37] - 2022-10-14
 
 ### Fixed